+++ packages/gasket-plugin-data/lib/actions.js
@@ -1,12 +1,11 @@
 /// <reference types="@gasket/core" />
 
-const { withGasketRequestCache } = require('@gasket/request');
 const deepClone = json => JSON.parse(JSON.stringify(json));
 
 const baseDataMap = new WeakMap();
 const adjustedDataMap = new WeakMap();
 
-/** @type {import('@gasket/core').ActionHandler<'getGasketData'>} */
+/** @type {import('@gasket/core', { with: { "resolution-mode": "import" } }).ActionHandler<'getGasketData'>} */
 async function getGasketData(gasket) {
   const key = gasket.symbol;
   if (!adjustedDataMap.has(key)) {
@@ -26,27 +25,26 @@ async function getGasketData(gasket) {
   return adjustedDataMap.get(key);
 }
 
-/** @type {import('@gasket/core').ActionHandler<'getPublicGasketData'>} */
-const getPublicGasketData = withGasketRequestCache(
-  async function getPublicGasketData(gasket, req) {
-    const basePublicData = (await gasket.actions.getGasketData()).public ?? {};
+/** @type {import('@gasket/core', { with: { "resolution-mode": "import" } }).ActionHandler<'getPublicGasketData'>} */
+async function getPublicGasketData(gasket, req) {
+  const basePublicData = (await gasket.actions.getGasketData()).public ?? {};
 
-    const userPublicData = await gasket.execWaterfall(
-      'publicGasketData',
-      deepClone(basePublicData),
-      { req }
+  const userPublicData = await gasket.execWaterfall(
+    'publicGasketData',
+    deepClone(basePublicData),
+    { req }
+  );
+
+  if (!userPublicData) {
+    throw new Error(
+      'No public gasketData - likely a publicGasketData lifecycle hook did not properly return data.'
     );
-
-    if (!userPublicData) {
-      throw new Error(
-        'No public gasketData - likely a publicGasketData lifecycle hook did not properly return data.'
-      );
-    }
-
-    return userPublicData;
   }
-);
 
+  return userPublicData;
+}
+
+// Export the base functions for testing
 module.exports = {
   baseDataMap,
   actions: {