+++ packages/gasket-plugin-nextjs/lib/fastify.js
@@ -2,37 +2,35 @@
 
 const { setupNextApp, setupNextHandling } = require('./utils/setup-next-app');
 
+/** Adds the buildId to Fastify via `decorate` */
+function registerBuildId(fastifyApp, app) {
+  const buildIdKey = ['buildId', app.name].filter(Boolean).join('/');
+  fastifyApp.decorate(buildIdKey, app.buildId);
+}
+
+/** Adds the `onResponse` hook to set NEXT_LOCALE */
+function addNextLocaleHook(fastifyApp) {
+  fastifyApp.addHook('onResponse', function setNextLocale(req, res, next) {
+    const locale = res?.locals?.gasketData?.intl?.locale;
+    if (locale) {
+      req.headers.cookie = (req.headers.cookie || '') + `;NEXT_LOCALE=${locale}`;
+    }
+    next();
+  });
+}
+
+/** @type {import('@gasket/core').HookHandler<'fastify'>} */
+async function fastifyHandler(gasket, fastifyApp) {
+  const app = await setupNextApp(gasket);
+  registerBuildId(fastifyApp, app);
+
+  await gasket.exec('nextFastify', { next: app, fastify: fastifyApp });
+
+  addNextLocaleHook(fastifyApp);
+  setupNextHandling(app, fastifyApp, gasket);
+}
+
 module.exports = {
-  timing: {
-    last: true
-  },
-  /** @type {import('@gasket/core').HookHandler<'fastify'>} */
-  handler: async function fastify(gasket, fastifyApp) {
-    const { exec } = gasket;
-
-    const app = await setupNextApp(gasket);
-
-    fastifyApp.decorate(
-      ['buildId', app.name].filter(Boolean).join('/'), app.buildId
-    );
-
-    await exec('nextFastify', { next: app, fastify: fastifyApp });
-
-    // TODO: Evaluate fix for this in Fastify4
-    fastifyApp.addHook('onResponse', function setNextLocale(req, res, next) {
-      // @ts-ignore
-      if (res.locals && res.locals.gasketData && res.locals.gasketData.intl) {
-        // @ts-ignore
-        const { locale } = res.locals.gasketData.intl;
-
-        if (locale) {
-          req.headers.cookie =
-            (req.headers.cookie || '') + `;NEXT_LOCALE=${locale}`;
-        }
-      }
-      next();
-    });
-
-    setupNextHandling(app, fastifyApp, gasket);
-  }
+  timing: { last: true },
+  handler: fastifyHandler
 };