+++ packages/gasket-plugin-winston/test/index.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-sync */
-const { Gasket } = require('@gasket/core');
+const { makeGasket } = require('@gasket/core');
 const PluginLogger = require('@gasket/plugin-logger');
 const plugin = require('../lib/index');
 const { LEVEL, MESSAGE } = require('triple-beam');
@@ -13,7 +13,7 @@ describe('@gasket/plugin-winston', function () {
   let gasket;
 
   beforeEach(() => {
-    gasket = new Gasket({ plugins: [PluginLogger, plugin] });
+    gasket = makeGasket({ plugins: [PluginLogger, plugin] });
     gasket.config = { env: 'local' };
   });