+++ scripts/generate-docs-index/CHANGELOG.md
@@ -1,5 +1,60 @@
 # generate-docs-index
 
+## 7.3.10
+
+### Patch Changes
+
+- Updated dependencies [a490392]
+- Updated dependencies [a490392]
+  - @gasket/plugin-nextjs@7.4.2
+  - @gasket/plugin-docusaurus@7.4.1
+  - @gasket/plugin-intl@7.4.0
+  - @gasket/preset-nextjs@7.3.8
+  - create-gasket-app@7.3.6
+  - @gasket/preset-api@7.3.8
+  - @gasket/plugin-analyze@7.3.4
+  - @gasket/plugin-command@7.3.5
+  - @gasket/plugin-cypress@7.4.0
+  - @gasket/plugin-data@7.3.3
+  - @gasket/plugin-docs@7.3.5
+  - @gasket/plugin-docs-graphs@7.3.3
+  - @gasket/plugin-dynamic-plugins@7.3.4
+  - @gasket/plugin-elastic-apm@7.3.3
+  - @gasket/plugin-express@7.3.4
+  - @gasket/plugin-fastify@7.3.5
+  - @gasket/plugin-git@7.3.3
+  - @gasket/plugin-happyfeet@7.3.3
+  - @gasket/plugin-https@7.3.4
+  - @gasket/plugin-https-proxy@7.3.4
+  - @gasket/plugin-jest@7.4.1
+  - @gasket/plugin-lint@7.3.5
+  - @gasket/plugin-logger@7.3.3
+  - @gasket/plugin-manifest@7.3.3
+  - @gasket/plugin-metadata@7.3.6
+  - @gasket/plugin-middleware@7.3.5
+  - @gasket/plugin-mocha@7.4.0
+  - @gasket/plugin-morgan@7.3.3
+  - @gasket/plugin-redux@7.4.0
+  - @gasket/plugin-service-worker@7.4.0
+  - @gasket/plugin-swagger@7.3.4
+  - @gasket/plugin-typescript@7.3.3
+  - @gasket/plugin-webpack@7.3.3
+  - @gasket/plugin-winston@7.3.3
+  - @gasket/plugin-workbox@7.3.3
+
+## 7.3.9
+
+### Patch Changes
+
+- Updated dependencies [d5ee612]
+- Updated dependencies [ea983c6]
+  - @gasket/plugin-nextjs@7.4.1
+  - @gasket/plugin-jest@7.4.1
+  - @gasket/nextjs@7.4.1
+  - @gasket/plugin-intl@7.4.0
+  - @gasket/preset-nextjs@7.3.7
+  - @gasket/preset-api@7.3.7
+
 ## 7.3.8
 
 ### Patch Changes